+++ frontend/pages/about.vue
@@ -1,0 +1,24 @@
+<template>
+  <Title>About the project</Title>
+  <section class="section about">
+    <div class="container is-widescreen">
+      <h1 class="title is-size-1">About this project</h1>
+      <p>
+        The idea behind this project was to create a web app to collect, classify and store links to
+        various useful web resources.
+      </p>
+      <p>
+        Write more about the project...
+      </p>
+      <p>
+        Link to the sources on GitHub.
+      </p>
+    </div>
+  </section>
+</template>
+
+<style scoped>
+.about {
+  min-height: calc(100vh - 52px - 168px);
+}
+</style>